♻️refactor: reconstruct file colocation
