+++ app/components/layout/footer.tsx
