+++ app/components/layout/heading.tsx
@@ -1,4 +1,4 @@
-import { Breadcrumb } from "@/app/components/breadcrumb"
+import { Breadcrumb } from "@/app/components/layout/breadcrumb"
 import type { Navigation } from "@/app/interfaces/navigation"
 import { navigation } from "@/app/lib/constant"
 import type React from "react"
