+++ app/globals.css
@@ -19,7 +19,7 @@
 }
 
 .typing {
-  animation: typing 2s steps(20), blink .7s infinite;
+  animation: typing 4s steps(20) infinite, blink .7s infinite;
   border-right: .15em solid orange;
   margin: auto;
   overflow: hidden;
@@ -90,9 +90,15 @@
 
 @keyframes typing {
   0% {
-    width: 0%;
+    width: 0;
+  }
+  20% {
+    width: 100%;
+  }
+  80% {
+    width: 100%;
   }
   100% {
-    width: 100%;
+    width: 0;
   }
 }
